chore(node_stream): split tests file under 2000-line gate
`crates/perry-runtime/src/node_stream_tests.rs` grew to 2039 lines on
this branch, tripping `scripts/check_file_size.sh` in the lint job.
Move the second half of the suite (54 tests starting at
`stream_max_listeners_default_and_override_round_trip`) into a sibling
`node_stream_tests_extra.rs`, mounted alongside `tests` as a separate
`#[cfg(test)] mod tests_extra` inside `node_stream`.

The shared capture state (thread-locals, extern "C" listener closures,
`string_value` helper) stays in `tests` and is now `pub(super)`; the
extra module pulls those fixtures in via an explicit named-import list
to avoid a name collision with `node_stream::string_value(&[u8])`,
which `use super::*` would otherwise drag in alongside `tests`'s
`string_value(&str)`.

No behavior change. `cargo fmt --all -- --check` and
`./scripts/check_file_size.sh` are clean; the 54 relocated tests still
pass under `cargo test -p perry-runtime --lib node_stream::`.
